+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix Computer/User status to display the correct values
 - Fix diagram size not respecting document border size
+- Fix in report only discovering child domains
 
 ## [0.9.9] - 2026-01-16